requirements.txt
gensim==3.6.0
numpy==1.17.3
pyemd==0.5.1
python-dotenv==0.10.3
scipy==1.1.0